requirements.txt
google-adk>=1.18.0
python-dotenv>=1.0.0

